extracted util functions
