+++ apps/verification-portal/src/lib/utils.ts
@@ -149,3 +149,14 @@ export function isFeatureEnabled(feature: string) {
 
   return isEnabled;
 }
+
+export function getCookie(name) {
+  const value = `; ${document.cookie}`;
+  const parts = value.split(`; ${name}=`);
+  if (parts.length === 2) return parts.pop().split(';').shift();
+}
+export function setCookie(name, value) {
+  const date = new Date();
+  date.setMonth(date.getMonth() + 1); // Set the cookie to expire in 1 month
+  document.cookie = `${name}=${value}; expires=${date.toUTCString()};path=/`;
+}
